Revert "【WIP】first_module_component_create"
