--- src/App.css
@@ -1,3 +1,0 @@
-body {
-  font-family: 'Open Sans Condensed';
-}
