+++ src/App.js
@@ -1,7 +1,6 @@
 import React from 'react';
 import './App.css';
-import HomePage from './pages/homepage/homepage.component';
-
+import HomePage from './homepage.component';
 
 function App() {
   return (
