--- src/pages/homepage/homepage.component.jsx
@@ -1,12 +1,0 @@
-import React from 'react';
-
-import Directory from '../../components/directory/directory.component';
-import './homepage.styles.scss';
-
-const HomePage = () => (
-  <div className='homepage'>
-    <Directory />
-  </div>
-);
-
-export default HomePage;
